Create base field unit test class.
- derived from unit test of FieldFE.
